requirements.txt
django==4.2
flake8==7.0.0
pytest-django==4.8.0
pytest-cov==7.0.0